+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-04-16 22:36:13
-// Commit: feat(perf): implement responsive perf view component
+// Updated at: 2026-04-17 09:10:19
+// Commit: chore(perf): update dependencies and cleanup unused imports in perf
 
 class perfComponentState {
-  final String commitRef = '447';
-  final String lastMessage = 'feat(perf): implement responsive perf view component';
-  final DateTime timestamp = DateTime.parse('2026-04-16T22:36:13+05:00');
+  final String commitRef = '451';
+  final String lastMessage = 'chore(perf): update dependencies and cleanup unused imports in perf';
+  final DateTime timestamp = DateTime.parse('2026-04-17T09:10:19+05:00');
 }